requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
braintree==4.15.2
Brotli==1.0.9
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cryptography==37.0.4
cssselect2==0.6.0
defusedxml==0.7.1
Django==4.0.5
django-sslserver==0.22
djangorestframework==3.13.1
dnspython==2.2.1
eventlet==0.33.1
flower==1.1.0
fonttools==4.33.3
greenlet==1.1.2
gunicorn==20.1.0
html5lib==1.1
humanize==4.2.3
idna==3.3
kombu==5.2.4
oauthlib==3.2.0
Pillow==9.1.1
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3

pycparser==2.21
pydyf==0.2.0
PyJWT==2.4.0
pyphen==0.12.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tinycss2==1.1.1
tornado==6.2
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
weasyprint==55.0
webencodings==0.5.1
whitenoise==6.2.0
zopfli==0.2.1
